Split actor data into multiple json files.
This has several advantages. First, it's much easier to find individual
monsters when they're organized into subcategories than it is when they
all live in one massive unorganized file. Second, adding appearance
randomization is going to be a snap under this new method. Third,
modding support is going to be much easier to achieve in the future.